#26 student can see data about himself: delete: SpecialityDTO; refactor: package "info-about-student" to "student", class InformationAboutStudentController to StudentController, class InformationAboutStudentService to StudentService; in V20200603130550__create-basic-database-structure.sql
