--- src/main/java/ua/edu/chdtu/deanoffice/mobile/backend/infoAboutStudent/dto/SpecialityDTO.java
@@ -1,10 +1,0 @@
-package ua.edu.chdtu.deanoffice.mobile.backend.infoAboutStudent.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class SpecialityDTO {
-    private String name;
-}
