+++ src/main/java/ua/edu/chdtu/deanoffice/mobile/backend/student/StudentController.java
@@ -1,23 +1,23 @@
-package ua.edu.chdtu.deanoffice.mobile.backend.infoAboutStudent;
+package ua.edu.chdtu.deanoffice.mobile.backend.student;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
-import ua.edu.chdtu.deanoffice.mobile.backend.infoAboutStudent.dto.StudentDTO;
+import ua.edu.chdtu.deanoffice.mobile.backend.student.dto.StudentDTO;
 
 import static ua.edu.chdtu.deanoffice.mobile.backend.general.mapper.Mapper.map;
 
 @RestController
-public class InformationAboutStudentController {
-    private InformationAboutStudentService restService;
+public class StudentController {
+    private StudentService restService;
 
     @Autowired
-    public InformationAboutStudentController(InformationAboutStudentService restService) {
+    public StudentController(StudentService restService) {
         this.restService = restService;
     }
 
-    @GetMapping("/info-about-student")
+    @GetMapping("/students")
     public ResponseEntity getInfoAboutStudent() {
         return ResponseEntity.ok(map(restService.getStudentInfo(), StudentDTO.class));
     }
